# openapi/beecount-cloud-v1.yaml
openapi: 3.1.0
info:
  title: BeeCount Cloud
  description: BeeCount Cloud v1 API
  version: 0.1.0
paths:
...
  /api/v1/sync/push:
    post:
      tags:
      - sync
      summary: Push Changes
      operationId: push_changes_api_v1_sync_push_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPushRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPushResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
  /api/v1/sync/pull:
    get:
      tags:
      - sync
      summary: Pull Changes
      operationId: pull_changes_api_v1_sync_pull_get
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      parameters:
      - name: since
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Since
      - name: device_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 5000
          minimum: 1
          default: 1000
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPullResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/full:
    get:
      tags:
      - sync
      summary: Full Snapshot
      operationId: full_snapshot_api_v1_sync_full_get
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      parameters:
      - name: ledger_id
        in: query
        required: true
        schema:
          type: string
          title: Ledger Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncFullResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/ledgers:
    get:
      tags:
      - sync
      summary: List Ledgers
      operationId: list_ledgers_api_v1_sync_ledgers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SyncLedgerOut'
                type: array
                title: Response List Ledgers Api V1 Sync Ledgers Get
      security:
      - OAuth2PasswordBearer:
        - app_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - web_write
      - OAuth2PasswordBearer:
        - ops_write
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SyncChangeIn:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        entity_type:
          type: string
          title: Entity Type
        entity_sync_id:
          type: string
          title: Entity Sync Id
        action:
          type: string
          enum:
          - upsert
          - delete
          title: Action
        payload:
          additionalProperties: true
          type: object
          title: Payload
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - ledger_id
      - entity_type
      - entity_sync_id
      - action
      - payload
      - updated_at
      title: SyncChangeIn
    SyncChangeOut:
      properties:
        change_id:
          type: integer
          title: Change Id
        ledger_id:
          type: string
          title: Ledger Id
        entity_type:
          type: string
          title: Entity Type
        entity_sync_id:
          type: string
          title: Entity Sync Id
        action:
          type: string
          enum:
          - upsert
          - delete
          title: Action
        payload:
          additionalProperties: true
          type: object
          title: Payload
        updated_at:
          type: string
          format: date-time
          title: Updated At
        updated_by_device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated By Device Id
      type: object
      required:
      - change_id
      - ledger_id
      - entity_type
      - entity_sync_id
      - action
      - payload
      - updated_at
      - updated_by_device_id
      title: SyncChangeOut
    SyncFullResponse:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        snapshot:
          anyOf:
          - $ref: '#/components/schemas/SyncChangeOut'
          - type: 'null'
        latest_cursor:
          type: integer
          title: Latest Cursor
      type: object
      required:
      - ledger_id
      - snapshot
      - latest_cursor
      title: SyncFullResponse
    SyncLedgerOut:
      properties:
        ledger_id:
          type: string
          title: Ledger Id
        path:
          type: string
          title: Path
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        size:
          type: integer
          title: Size
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        role:
          type: string
          enum:
          - owner
          - editor
          - viewer
          title: Role
      type: object
      required:
      - ledger_id
      - path
      - updated_at
      - size
      - metadata
      - role
      title: SyncLedgerOut
    SyncPullResponse:
      properties:
        changes:
          items:
            $ref: '#/components/schemas/SyncChangeOut'
          type: array
          title: Changes
        server_cursor:
          type: integer
          title: Server Cursor
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - changes
      - server_cursor
      - has_more
      title: SyncPullResponse
    SyncPushRequest:
      properties:
        device_id:
          type: string
          title: Device Id
        changes:
          items:
            $ref: '#/components/schemas/SyncChangeIn'
          type: array
          title: Changes
      type: object
      required:
      - device_id
      - changes
      title: SyncPushRequest
    SyncPushResponse:
      properties:
        accepted:
          type: integer
          title: Accepted
        rejected:
          type: integer
          title: Rejected
        conflict_count:
          type: integer
          title: Conflict Count
          default: 0
        conflict_samples:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Conflict Samples
        server_cursor:
          type: integer
          title: Server Cursor
        server_timestamp:
          type: string
          format: date-time
          title: Server Timestamp
      type: object
      required:
      - accepted
      - rejected
      - server_cursor
      - server_timestamp
      title: SyncPushResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes:
            app_write: App client read/write scope
            web_read: Web console read scope
            web_write: Web console write scope
            ops_write: Operations and admin scope
          tokenUrl: /api/v1/auth/login